Balance fix

+++ Assets/scripts/Core/Runtime/Interfaces/IHealthSystem.cs
@@ -3,9 +3,9 @@
     public interface IHealthSystem : IEntity
     {
         void Init(EntityType type);
-        int MaxHealth { get; }
-        int CurrentHealth { get; }
-        void ModifyHealth(int amount); //implementations must support both positive and negative amounts
+        float MaxHealth { get; }
+        float CurrentHealth { get; }
+        void ModifyHealth(float amount); //implementations must support both positive and negative amounts
         void Kill(); //what happens when current health is <= 0
         bool IsDead { get; }
     }
